By proper use of generics, a bunch of casts are not needed.
git-svn-id: https://svn.apache.org/repos/asf/cxf/trunk@1228848 13f79535-47bb-0310-9956-ffa450edef68

+++ rt/bindings/corba/src/main/java/org/apache/cxf/binding/corba/utils/CorbaAnyHelper.java
@@ -202,19 +202,19 @@ public final class CorbaAnyHelper {
             primitive.setValue(Short.valueOf(a.extract_short()));
             break;
         case TCKind._tk_ushort:
-            primitive.setValue(Integer.valueOf((int)a.extract_ushort()));
+            primitive.setValue(Integer.valueOf(a.extract_ushort()));
             break;
         case TCKind._tk_long:
             primitive.setValue(Integer.valueOf(a.extract_long()));
             break;
         case TCKind._tk_ulong:
-            primitive.setValue(BigInteger.valueOf((long)a.extract_ulong()));
+            primitive.setValue(BigInteger.valueOf(a.extract_ulong()));
             break;
         case TCKind._tk_longlong:
             primitive.setValue(Long.valueOf(a.extract_longlong()));
             break;
         case TCKind._tk_ulonglong:
-            primitive.setValue(BigInteger.valueOf((long)a.extract_ulonglong()));
+            primitive.setValue(BigInteger.valueOf(a.extract_ulonglong()));
             break;
         case TCKind._tk_float:
             primitive.setValue(Float.valueOf(a.extract_float()));
